feat: added vertexConstantAttributeBuilder
