--- packages/d3fc-webgl/src/buffers/projectedAttributeBuilder.js
@@ -1,69 +1,0 @@
-import baseAttributeBuilder from './baseAttributeBuilder';
-import { rebindAll, exclude } from '@d3fc/d3fc-rebind';
-
-export default () => {
-    const base = baseAttributeBuilder();
-
-    let value = (data, element, vertex, component, index) => data[index];
-    let data = null;
-
-    const project = (elementCount, verticesPerElement) => {
-        const components = base.size();
-        const offset = base.offset();
-        const projectedData = new Float32Array(
-            offset + elementCount * verticesPerElement * components
-        );
-        let index = 0;
-        for (let element = 0; element < elementCount; element++) {
-            for (let vertex = 0; vertex < verticesPerElement; vertex++) {
-                for (let component = 0; component < components; component++) {
-                    projectedData[offset + index++] = value(
-                        data,
-                        element,
-                        vertex,
-                        component,
-                        index
-                    );
-                }
-            }
-        }
-        return projectedData;
-    };
-
-    const build = (gl, program, name, verticesPerElement, count) => {
-        base(gl, program, name);
-
-        if (base.validSize() >= count) {
-            return;
-        }
-
-        const projectedData = project(count, verticesPerElement);
-
-        gl.bindBuffer(gl.ARRAY_BUFFER, base.buffer());
-        gl.bufferData(gl.ARRAY_BUFFER, projectedData, gl.DYNAMIC_DRAW);
-
-        base.validSize(count);
-    };
-
-    build.value = (...args) => {
-        if (!args.length) {
-            return value;
-        }
-        value = args[0];
-        base.validSize(0);
-        return build;
-    };
-
-    build.data = (...args) => {
-        if (!args.length) {
-            return data;
-        }
-        data = args[0];
-        base.validSize(0);
-        return build;
-    };
-
-    rebindAll(build, base, exclude('buffer'));
-
-    return build;
-};
